Crate zink_abi_macro

Macros

import

The import! macro generates a Rust struct and implementation for interacting with an Ethereum smart contract based on its ABI (Application Binary Interface) and deploys the corresponding contract.

Structs

AbiFunction 🔒

A struct to represent the function in an ERC ABI

AbiParameter 🔒

A struct to represent a parameter in an ERC ABI

EthereumAbi 🔒

Represents an Ethereum ABI

Functions

generate_function_implementation 🔒

Generate the implementation for a contract function

generate_function_signature 🔒

Generate a function signature for an ABI function

map_type_to_rust_and_encode 🔒

Maps Solidity types to Rust types and handles encoding/decoding